requirements.txt
attrs==25.3.0
black==25.1.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
docopt==0.6.2
docstring-to-markdown==0.15
filelock==3.18.0
idna==3.10
jedi==0.19.2
mypy_extensions==1.0.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
pycparser==2.22
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.4
ujson==5.10.0
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1

# Añadidos mínimos para FastAPI Starter
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

python-jose==3.3.0
pytest==8.2.2
httpx==0.27.0

python-multipart

SQLAlchemy>=2.0
alembic>=1.13
passlib[bcrypt]>=1.7.4

email-validator>=2.1.0
bcrypt==4.0.1
